make vector_graphics_compiler compile on the web again
