+++ packages/vector_graphics_compiler/lib/src/svg/path_ops.dart
@@ -6,7 +6,7 @@
 
 import '_path_ops_unsupported.dart' if (dart.library.ffi) '_path_ops_ffi.dart'
     as impl;
-export '_path_ops_ffi.dart';
+export '_path_ops_unsupported.dart' if (dart.library.ffi) '_path_ops_ffi.dart';
 
 // ignore_for_file: camel_case_types, non_constant_identifier_names
 
@@ -88,4 +88,38 @@ abstract class PathProxy {
 
   /// Called by [Path.replay] to indicate that a new path is being played.
   void reset() {}
+}
+
+/// A path proxy that can print the SVG path-data representation of this path.
+class SvgPathProxy implements PathProxy {
+  final StringBuffer _buffer = StringBuffer();
+
+  @override
+  void reset() {
+    _buffer.clear();
+  }
+
+  @override
+  void close() {
+    _buffer.write('Z');
+  }
+
+  @override
+  void cubicTo(
+      double x1, double y1, double x2, double y2, double x3, double y3) {
+    _buffer.write('C$x1,$y1 $x2,$y2 $x3,$y3');
+  }
+
+  @override
+  void lineTo(double x, double y) {
+    _buffer.write('L$x,$y');
+  }
+
+  @override
+  void moveTo(double x, double y) {
+    _buffer.write('M$x,$y');
+  }
+
+  @override
+  String toString() => _buffer.toString();
 }
